Split censor, pings, command checks into separate Cogs and files
